refactor: consolidate portfolio management into hooks and remove context
