+++ frontend/src/App.tsx
@@ -1,14 +1,8 @@
-import { PortfolioPageProvider } from "@context/portfolioPageContext";
 import "./App.css";
 import PortfolioPage from "@pages/PortfolioPage";
 
 function App() {
-  console.log("in App!!");
-  return (
-    <PortfolioPageProvider>
-      <PortfolioPage />
-    </PortfolioPageProvider>
-  );
+  return <PortfolioPage />;
 }
 
 export default App;
