+++ frontend/src/components/AddAssetModal.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useRef } from "react";
 import { AssetSelector } from "@components/AssetSelector";
 import { FocusTrap } from "focus-trap-react";
-import type { PortfolioAsset } from "@hooks/usePortfolio";
+import type { PortfolioAsset } from "@hooks/useUIState";
 import { useAddAssetForm } from "@hooks/useAddAssetForm";
 
 type Props = {
