+++ frontend/src/hooks/usePortfolioState.ts
@@ -1,29 +1,45 @@
-import { useState, useRef } from "react";
-import { usePortfolio } from "@hooks/usePortfolio";
+import { useState, useCallback } from "react";
+export type PortfolioAsset = {
+  id: string; // e.g., "bitcoin"
+  symbol: string; // e.g., "BTC"
+  name: string; // e.g., "Bitcoin"
+  quantity: number;
+};
+
+export type PortfolioState = PortfolioAsset[];
+
 /**
  * Hook to manage portfolio list, modal state, and add-button focus.
  */
 export function usePortfolioState() {
-  const { portfolio, addAsset } = usePortfolio();
+  const [portfolio, setPortfolio] = useState<PortfolioState>([]);
 
-  const [showModal, setShowModal] = useState(false);
-  const addButtonRef = useRef<HTMLButtonElement | null>(null);
+  const addAsset = useCallback((newAsset: PortfolioAsset) => {
+    setPortfolio((prev) => {
+      const existing = prev.find((asset) => asset.id === newAsset.id);
+      if (existing) {
+        return prev.map((asset) =>
+          asset.id === newAsset.id
+            ? { ...asset, quantity: asset.quantity + newAsset.quantity }
+            : asset
+        );
+      }
+      return [...prev, newAsset];
+    });
+  }, []);
 
-  const openModal = () => setShowModal(true);
+  const removeAsset = useCallback((assetId: string) => {
+    setPortfolio((prev) => prev.filter((asset) => asset.id !== assetId));
+  }, []);
 
-  const closeModal = () => {
-    setShowModal(false);
-    setTimeout(() => {
-      addButtonRef.current?.focus();
-    }, 0); // ensure modal fully unmounts
-  };
+  const resetPortfolio = useCallback(() => {
+    setPortfolio([]);
+  }, []);
 
   return {
     portfolio,
     addAsset,
-    showModal,
-    openModal,
-    closeModal,
-    addButtonRef,
+    removeAsset,
+    resetPortfolio,
   };
 }
